requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram.git@v2.1.13
tgcrypto
aiohttp<=3.8.1
python-dotenv<=0.20.0
motor
aiofiles
dnspython
apscheduler
requests
psutil
pyromod
colorama
cachetools
